Detection/requirements.txt
gradio==3.50.2
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
tensorflow==2.16.1
torch==2.1.0
torchvision==0.16.0
albumentations==1.3.1